[python-frontend] restore function call collection handling to fix python regression failures
